Refactor(#16) : 마이페이지 속 인증 패키지 생성해 파일 이동
