+++ src/main/java/com/keodam/keodam_backend/mypage/certification/domain/DocumentType.java
@@ -1,0 +1,5 @@
+package com.keodam.keodam_backend.mypage.certification.domain;
+
+public enum DocumentType {
+    ENROLLMENT, GRADUATION, EMPLOYMENT, OFFLINE_SHOT
+}
