+++ src/main/java/com/keodam/keodam_backend/mypage/certification/domain/VerificationStatus.java
@@ -1,0 +1,5 @@
+package com.keodam.keodam_backend.mypage.certification.domain;
+
+public enum VerificationStatus {
+    PENDING,APPROVED,REJECTED
+}
