+++ src/main/java/com/keodam/keodam_backend/mypage/certification/dto/request/FileRequestDto.java
@@ -1,0 +1,8 @@
+package com.keodam.keodam_backend.mypage.certification.dto.request;
+
+import lombok.Getter;
+
+@Getter
+public class FileRequestDto {
+    private String documentType;
+}
