--- src/main/java/com/keodam/keodam_backend/mypage/domain/DocumentType.java
@@ -1,5 +1,0 @@
-package com.keodam.keodam_backend.mypage.domain;
-
-public enum DocumentType {
-    ENROLLMENT, GRADUATION, EMPLOYMENT, OFFLINE_SHOT
-}
