--- src/main/java/com/keodam/keodam_backend/mypage/domain/VerificationStatus.java
@@ -1,5 +1,0 @@
-package com.keodam.keodam_backend.mypage.domain;
-
-public enum VerificationStatus {
-    PENDING,APPROVED,REJECTED
-}
