--- src/main/java/com/keodam/keodam_backend/mypage/dto/request/FileRequestDto.java
@@ -1,8 +1,0 @@
-package com.keodam.keodam_backend.mypage.dto.request;
-
-import lombok.Getter;
-
-@Getter
-public class FileRequestDto {
-    private String documentType;
-}
