--- src/main/java/com/keodam/keodam_backend/mypage/dto/response/EmailResponseDto.java
@@ -1,17 +1,0 @@
-package com.keodam.keodam_backend.mypage.dto.response;
-
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-public class EmailResponseDto {
-
-    private String code;
-
-    @Builder
-    public EmailResponseDto(String code) {
-        this.code = code;
-    }
-}
